requirements.txt
streamlit>=1.42.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
pypdf>=5.0.0
python-dotenv>=1.0.1
pandas>=2.2.0
sentence-transformers>=3.0.0
requests>=2.32.0